Move Python task helpers to internal/python
