+++ internal/helpers.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"path"
 	"path/filepath"
-	"strings"
 
 	"github.com/databricks/cli/internal/acc"
 	"github.com/databricks/cli/internal/testutil"
@@ -17,79 +16,10 @@ import (
 	"github.com/databricks/databricks-sdk-go"
 	"github.com/databricks/databricks-sdk-go/apierr"
 	"github.com/databricks/databricks-sdk-go/service/catalog"
-	"github.com/databricks/databricks-sdk-go/service/compute"
 	"github.com/databricks/databricks-sdk-go/service/files"
-	"github.com/databricks/databricks-sdk-go/service/jobs"
 	"github.com/databricks/databricks-sdk-go/service/workspace"
 	"github.com/stretchr/testify/require"
 )
-
-func GenerateNotebookTasks(notebookPath string, versions []string, nodeTypeId string) []jobs.SubmitTask {
-	tasks := make([]jobs.SubmitTask, 0)
-	for i := 0; i < len(versions); i++ {
-		task := jobs.SubmitTask{
-			TaskKey: fmt.Sprintf("notebook_%s", strings.ReplaceAll(versions[i], ".", "_")),
-			NotebookTask: &jobs.NotebookTask{
-				NotebookPath: notebookPath,
-			},
-			NewCluster: &compute.ClusterSpec{
-				SparkVersion:     versions[i],
-				NumWorkers:       1,
-				NodeTypeId:       nodeTypeId,
-				DataSecurityMode: compute.DataSecurityModeUserIsolation,
-			},
-		}
-		tasks = append(tasks, task)
-	}
-
-	return tasks
-}
-
-func GenerateSparkPythonTasks(notebookPath string, versions []string, nodeTypeId string) []jobs.SubmitTask {
-	tasks := make([]jobs.SubmitTask, 0)
-	for i := 0; i < len(versions); i++ {
-		task := jobs.SubmitTask{
-			TaskKey: fmt.Sprintf("spark_%s", strings.ReplaceAll(versions[i], ".", "_")),
-			SparkPythonTask: &jobs.SparkPythonTask{
-				PythonFile: notebookPath,
-			},
-			NewCluster: &compute.ClusterSpec{
-				SparkVersion:     versions[i],
-				NumWorkers:       1,
-				NodeTypeId:       nodeTypeId,
-				DataSecurityMode: compute.DataSecurityModeUserIsolation,
-			},
-		}
-		tasks = append(tasks, task)
-	}
-
-	return tasks
-}
-
-func GenerateWheelTasks(wheelPath string, versions []string, nodeTypeId string) []jobs.SubmitTask {
-	tasks := make([]jobs.SubmitTask, 0)
-	for i := 0; i < len(versions); i++ {
-		task := jobs.SubmitTask{
-			TaskKey: fmt.Sprintf("whl_%s", strings.ReplaceAll(versions[i], ".", "_")),
-			PythonWheelTask: &jobs.PythonWheelTask{
-				PackageName: "my_test_code",
-				EntryPoint:  "run",
-			},
-			NewCluster: &compute.ClusterSpec{
-				SparkVersion:     versions[i],
-				NumWorkers:       1,
-				NodeTypeId:       nodeTypeId,
-				DataSecurityMode: compute.DataSecurityModeUserIsolation,
-			},
-			Libraries: []compute.Library{
-				{Whl: wheelPath},
-			},
-		}
-		tasks = append(tasks, task)
-	}
-
-	return tasks
-}
 
 func TemporaryWorkspaceDir(t testutil.TestingT, w *databricks.WorkspaceClient) string {
 	ctx := context.Background()
